Refactored NPCs to derive from characters - this will give them all the same base stats, etc.
